refactor(namespace): refactor namespace package to vite-based build
