+++ src/common/namespace/internal/index.ts
@@ -1,0 +1,2 @@
+export { createNamespacePrepender, type Namespaced } from './src/createNamespacePrepender';
+export { isValidNamespace, type ValidNamespace } from './src/isValidNamespace';
