+++ src/common/namespace/internal/package.json
@@ -1,0 +1,28 @@
+{
+    "name": "@docodylus/namespace-internal",
+    "version": "0.0.1-prerelease.3",
+    "liceense": "Apache-2.0",
+    "type": "module",
+    "sideEffects": false,
+    "exports": {
+        ".": {
+            "import": "./dist/esm/index.mjs",
+            "require": "./dist/cjs/index.cjs",
+            "types": "./dist/index.d.ts"
+        }
+    },
+    "files": [
+        "dist/esm",
+        "dist/cjs",
+        "dist/index.d.ts"
+    ],
+    "scripts": {
+        "build": "../../../../node_modules/.bin/vite build",
+        "clean": "rm -rf dist",
+        "lint": "eslint . --ext .ts",
+        "test": "vitest run"
+    },
+    "dependencies": {
+        "@docodylus/validation-internal": "workspace:*"
+    }
+}
