+++ src/common/namespace/internal/vite.config.ts
@@ -1,0 +1,2 @@
+import { createInternalViteConfig} from '../../../../infra/build/packages/vite.config.base';
+export default createInternalViteConfig('docodylusValidationInternal');
