--- src/common/namespace/src/index.ts
@@ -1,3 +1,0 @@
-export { createNamespacePrepender } from './createNamespacePrepender';
-export { isValidNamespace } from './isValidNamespace';
-export type { Namespaced, ValidNamespace } from '@namespace/types';
