--- src/common/namespace/src/isValidNamespace.ts
@@ -1,6 +1,0 @@
-const namespaceRegex = /^([a-z]+\.)*[a-z]+$/;
-
-export function isValidNamespace(maybeNamespace: unknown): boolean {
-  if (typeof maybeNamespace !== 'string') return false;
-  return namespaceRegex.test(maybeNamespace);
-}
