--- src/common/namespace/src/types/index.ts
@@ -1,2 +1,0 @@
-export { Namespaced } from './Namespaced';
-export { ValidNamespace } from './ValidNamespace';
